Refactor model naming
