+++ OpenReferralApi/Controllers/StatusController.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel;
 using Microsoft.AspNetCore.Mvc;
 using OpenReferralApi.Models;
+using OpenReferralApi.Models.Responses;
 
 namespace OpenReferralApi.Controllers;
 
@@ -13,12 +14,12 @@ public class StatusController : ControllerBase
     /// A status endpoint that can be used to check if the API is running and healthy
     /// </summary>
     [HttpGet]
-    [ProducesResponseType(typeof(StatusResult), 200)]
-    [ProducesResponseType(typeof(StatusResult), 500)]
+    [ProducesResponseType(typeof(StatusResponse), 200)]
+    [ProducesResponseType(typeof(StatusResponse), 500)]
     [Description("A status endpoint that can be used to check if the API is running and healthy")]
     public IActionResult GetStatus()
     {
-        var statusResult = new StatusResult()
+        var statusResult = new StatusResponse()
         {
             Healthy = true,
             Message = "Service running"
