+++ OpenReferralApi/Models/Responses/DashboardValidationResponse.cs
@@ -1,4 +1,4 @@
-namespace OpenReferralApi.Models;
+namespace OpenReferralApi.Models.Responses;
 
 public class DashboardValidationResponse
 {
