+++ OpenReferralApi/Services/Interfaces/IDashboardService.cs
@@ -1,5 +1,7 @@
 using FluentResults;
 using OpenReferralApi.Models;
+using OpenReferralApi.Models.Requests;
+using OpenReferralApi.Models.Responses;
 
 namespace OpenReferralApi.Services.Interfaces;
 
@@ -8,5 +10,5 @@ public interface IDashboardService
     public Task<Result<DashboardOutput>> GetServices();
     public Task<Result<DashboardServiceDetails>> GetServiceById(string id);
     public Task<Result<List<DashboardValidationResponse>>> ValidateDashboardServices();
-    public Task<Result<SubmissionResponse>> SubmitService(DashboardSubmission submission);
+    public Task<Result<SubmissionResponse>> SubmitService(DashboardSubmissionRequest submission);
 }
