+++ OpenReferralApi/Services/Interfaces/IGithubService.cs
@@ -1,9 +1,11 @@
 using FluentResults;
 using OpenReferralApi.Models;
+using OpenReferralApi.Models.Requests;
+using OpenReferralApi.Models.Responses;
 
 namespace OpenReferralApi.Services.Interfaces;
 
 public interface IGithubService
 {
-    public Task<Result<SubmissionResponse>> RaiseIssue(DashboardSubmission submission);
+    public Task<Result<SubmissionResponse>> RaiseIssue(DashboardSubmissionRequest submission);
 }
