+++ OpenReferralApi/Services/Interfaces/IValidatorService.cs
@@ -1,6 +1,7 @@
 using FluentResults;
 using Json.Schema;
 using OpenReferralApi.Models;
+using OpenReferralApi.Models.Responses;
 
 namespace OpenReferralApi.Services.Interfaces;
